en cours...

+++ contact_app/src/main/java/isen/java2/model/db/entities/Category.java
@@ -22,7 +22,7 @@ public class Category {
 		return id;
 	}
 	
-	public void setId(short id) {
+	public void setId(int id) {
 		this.id = id;
 	}
